[web] Show `DetailModal` on asset click in `DetailModal`
